+++ qutebrowser/completion/completer.py
@@ -22,7 +22,7 @@
 from PyQt5.QtCore import pyqtSignal, pyqtSlot, QObject, QTimer
 
 from qutebrowser.config import config
-from qutebrowser.commands import cmdexc, cmdutils, runners
+from qutebrowser.commands import cmdutils, runners
 from qutebrowser.utils import usertypes, log, objreg, utils
 from qutebrowser.completion.models import instances, sortfilter
 
@@ -485,16 +485,3 @@ class Completer(QObject):
         """Select the next completion item."""
         self._open_completion_if_needed()
         self.next_prev_item.emit(False)
-
-    @cmdutils.register(instance='completion', hide=True,
-                       modes=[usertypes.KeyMode.command], scope='window')
-    def completion_item_del(self):
-        """Delete the current completion item."""
-        completion = objreg.get('completion', scope='window',
-                                window=self._win_id)
-        if not completion.currentIndex().isValid():
-            raise cmdexc.CommandError("No item selected!")
-        try:
-            self.model().srcmodel.delete_cur_item(completion)
-        except NotImplementedError:
-            raise cmdexc.CommandError("Cannot delete this item.")
